fix: move contexts down
work in progress
