+++ src/modules/common/Pausable.tsx
@@ -2,11 +2,10 @@ import { FC, ReactElement } from 'react';
 
 import { ActivityStatus } from '@/interfaces/interactionProcess';
 import WaitingScreen from '@/modules/common/WaitingScreen';
-
-import { useActivityContext } from '../context/ActivityContext';
+import { useActivityStateContext } from '@/modules/context/ActivityStateContext';
 
 const Pausable: FC<{ children: ReactElement }> = ({ children }) => {
-  const { activityState } = useActivityContext();
+  const { activityState } = useActivityStateContext();
   const status = activityState?.data?.status || ActivityStatus.WaitingForStart;
 
   if (status === ActivityStatus.Play) {
