+++ src/modules/common/ResetSetsButton.tsx
@@ -9,8 +9,9 @@ import DialogContentText from '@mui/material/DialogContentText';
 import DialogTitle from '@mui/material/DialogTitle';
 
 import usePrompts from '@/hooks/usePrompts';
+import { useActivityStateContext } from '@/modules/context/ActivityStateContext';
 
-import { useActivityContext } from '../context/ActivityContext';
+import { useResponses } from '../context/ResponsesContext';
 
 interface ResetSetsButtonProps {
   enable?: boolean;
@@ -21,7 +22,8 @@ const ResetSetsButton: FC<ResetSetsButtonProps> = (
 ) => {
   const { t } = useTranslation();
   const [dialogOpen, setDialogOpen] = useState(false);
-  const { deleteAllResponsesSet, resetActivityState } = useActivityContext();
+  const { resetActivityState } = useActivityStateContext();
+  const { deleteAllResponsesSet } = useResponses();
   const { resetAllPrompts } = usePrompts();
   const promise = useRef<Promise<void>>();
