+++ src/modules/orchestration/OrchestrationBar.tsx
@@ -10,8 +10,8 @@ import useTheme from '@mui/material/styles/useTheme';
 import { ORCHESTRATION_BAR_CY } from '@/config/selectors';
 import useSteps from '@/hooks/useSteps';
 import { ActivityStatus } from '@/interfaces/interactionProcess';
+import { useActivityStateContext } from '@/modules/context/ActivityStateContext';
 
-import { useActivityContext } from '../context/ActivityContext';
 import CommandButton from './CommandButton';
 import StepsButton from './StepsButton';
 
@@ -23,7 +23,8 @@ const OrchestrationBar: FC<OrchestrationBarProps> = () => {
   const { t } = useTranslation('translations', {
     keyPrefix: 'ORCHESTRATION_BAR',
   });
-  const { activityState, pauseActivity, playActivity } = useActivityContext();
+  const { activityState, pauseActivity, playActivity } =
+    useActivityStateContext();
   const { currentStep, nextStep } = useSteps();
   const { status } = activityState.data;
